Print RGB Bar Graphs as Guides for Hue, Saturation and Light

Print bar charts of RGB values that represent simplified groupings of hue, saturation or lightness levels in the HSL system. Intended for use as a reference to assess which RGB profiles most closely represents a user's hex shortcode.

dh_guide(type = c("H", "S", "L"), crayon = TRUE)

Arguments

type
Character. Which of hue ('H'), saturation ('S') or light ('L') that you want to print the guide for.

crayon
Logical. Do you want to print to the console in colour, using the crayon package?

Value

Nothing. Prints bar charts of RGB values to the console, made with unicode blocks.

Details

The hue guide prints guides for the primary (red, green, blue), secondary (yellow, cyan, magenta) and tertiary (orange, chartreuse, aquamarine, azure, violet, rose) hues. For saturation, the guide contains 'washed' (i.e. the largest range in RGB values), muted, and grey (i.e. no difference between RGB values). In the lightness guide they're 'light' (i.e.high mean value of RGB), 'middle', and 'dark' (i.e. small range in RGB values).

Examples

dh_guide("L")
#> light
#> R ████████████████
#> G ███████████████░
#> B ██████████████░░
#> L ░░░░░░░░░░░░░░█░
#> 
#> middle
#> R ██████████░░░░░░
#> G █████████░░░░░░░
#> B ████████░░░░░░░░
#> L ░░░░░░░░█░░░░░░░
#> 
#> dark
#> R ████░░░░░░░░░░░░
#> G ███░░░░░░░░░░░░░
#> B ██░░░░░░░░░░░░░░
#> L ░░█░░░░░░░░░░░░░
#>